requirements.txt
pytesseract
opencv-python
numpy
pillow
pyautogui
pydirectinput
pyinstaller
